# api/src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ApiResponse": {
				"properties": {
					"status": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"success": {
						"type": "boolean"
					},
					"data": {
						"properties": {},
						"type": "object",
						"nullable": true
					},
					"token": {
						"type": "string"
					},
					"code": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"code"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CreateParkingEntryDto": {
				"properties": {
					"plateNumber": {
						"type": "string"
					},
					"parkingId": {
						"type": "string"
					}
				},
				"required": [
					"plateNumber",
					"parkingId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ExitParkingDto": {
				"properties": {
					"entryId": {
						"type": "string"
					}
				},
				"required": [
					"entryId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "cielo-b"
		}
	},
	"paths": {
...
		"/parking-entry/entry": {
			"post": {
				"operationId": "CreateEntry",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					},
					"201": {
						"description": "Vehicle entry recorded successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request - No available spaces or vehicle already in parking",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					},
					"404": {
						"description": "Parking not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Parking Entry"
				],
				"security": [
					{
						"bearerAuth": [
							"ATTENDANT"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateParkingEntryDto"
							}
						}
					}
				}
			}
		},
		"/parking-entry/exit": {
			"post": {
				"operationId": "ProcessExit",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request - Vehicle already exited",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					},
					"404": {
						"description": "Parking entry not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Parking Entry"
				],
				"security": [
					{
						"bearerAuth": [
							"ATTENDANT"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ExitParkingDto"
							}
						}
					}
				}
			}
		},
		"/parking-entry/active": {
			"get": {
				"operationId": "GetActiveEntries",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Parking Entry"
				],
				"security": [
					{
						"bearerAuth": [
							"ATTENDANT",
							"ADMIN"
						]
					}
				],
				"parameters": []
			}
		},
		"/parking-entry/report": {
			"get": {
				"operationId": "GetEntriesByDateRange",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Parking Entry"
				],
				"security": [
					{
						"bearerAuth": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "startDate",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "endDate",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/parking-entry/ticket": {
			"post": {
				"operationId": "GenerateTicket",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					},
					"404": {
						"description": "Parking entry not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Parking Entry"
				],
				"security": [
					{
						"bearerAuth": [
							"ATTENDANT"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"entryId": {
										"type": "string"
									}
								},
								"required": [
									"entryId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/parking-entry/bills": {
			"get": {
				"operationId": "GetBills",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Parking Entry"
				],
				"security": [
					{
						"bearerAuth": [
							"ATTENDANT",
							"ADMIN"
						]
					}
				],
				"parameters": []
			}
		},
		"/parking-entry/bill/{id}": {
			"get": {
				"operationId": "GenerateBill",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					},
					"404": {
						"description": "Parking entry not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Parking Entry"
				],
				"security": [
					{
						"bearerAuth": [
							"ATTENDANT",
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}